Better kind support for real/comlpex calls
Make sure all routines use AI_kinds : DP
Make sure that calls to real( ) have an explicit kind
 Changes to be committed:
	modified:   Makefile
	modified:   OCEAN_bloch.f90
	modified:   OCEAN_haydock.f90
	modified:   OCEAN_hyb_louie_levine.f90
	modified:   OCEAN_invdrv.f90
	modified:   OCEAN_multiplet.f90
	modified:   OCEAN_obf.f90
	modified:   long_range.f90
	modified:   nbsemkcmel.f90
	modified:   sphsetnx.h.f90
	modified:   xes_spaghetti.f90

+++ OCEAN2/OCEAN_haydock.f90
@@ -1084,7 +1084,7 @@ module OCEAN_action
 
     integer :: dumi, iter, ierr_
     character(len=4) :: inv_style
-    real :: dumf
+    real( DP ) :: dumf
 
     if( .not. is_first ) goto 10
     is_first = .false.
